backend/services/staff_customer_classification_service/requirements.txt
fastai==2.8.7
torch==2.10.0
torchvision==0.25.0
fastcore==1.12.24
fasttransform==0.0.2
fastdownload==0.0.7

numpy
pillow
opencv-python
requests
python-dotenv
python-multipart
python-logging-loki
litserve==0.2.5
ipython
